fixed ui bug with edit and delete in combination with new responsehandling

+++ miraBeauWebApp/src/pages/groupTable/group-page.ts
@@ -113,10 +113,13 @@ export class groupPage {
     this.dialogService.open({
       viewModel: DeleteDialog,
       model: "Are you sure you want to delete this person?"
-    }).whenClosed(response => {
-      this.peopleApi.deletePerson(dude.id).then(() => {
-        this.getAllPeople();
-      });
+    }).whenClosed(result => {
+      if (!result.wasCancelled) {
+        this.peopleApi.deletePerson(dude.id).then(() => {
+          this.newlyAdded = false;
+          this.getAllPeople();
+        });
+      }
     });
   }
